+++ src/system/celestial_body/gaseous/generator.rs
@@ -15,12 +15,12 @@ impl GaseousBodyDetails {
         star_id: u32,
         star_name: Rc<str>,
         star_age: f32,
-        star_mass: f32,
+        star_mass: f64,
         star_type: &StarSpectralType,
         star_class: &StarLuminosityClass,
         star_luminosity: f32,
         star_traits: &Vec<StarPeculiarity>,
-        primary_star_mass: f32,
+        primary_star_mass: f64,
         gas_giant_arrangement: GasGiantArrangement,
         orbit: Orbit,
         orbit_distance: f64,
@@ -59,7 +59,7 @@ impl GaseousBodyDetails {
             Self::get_gas_body_size_modifier(star_mass, star_type, &mut rng, is_proto_giant);
         let roll_result = rng.roll(1, 400, size_modifier);
 
-        let mut mass: f32 = 0.0;
+        let mut mass: f64 = 0.0;
         let mut size = CelestialBodySize::Puny;
         let blackbody_temp = calculate_blackbody_temperature(star_luminosity, orbit_distance);
         if !is_proto_giant && roll_result <= 2 {
@@ -96,32 +96,32 @@ impl GaseousBodyDetails {
             ));
         } else if roll_result <= 106 {
             // Gas planet (2 to 16 masses)
-            mass = rng.roll(1, 1600 - (200 - 1), 200 - 1) as f32 / 100.0;
+            mass = rng.roll(1, 1600 - (200 - 1), 200 - 1) as f64 / 100.0;
             size = CelestialBodySize::Large;
         } else if roll_result <= 326 {
             // Gas giant (16.001 to 162)
-            mass = rng.roll(1, 16200 - (1600 - 1), 1600 - 1) as f32 / 100.0;
+            mass = rng.roll(1, 16200 - (1600 - 1), 1600 - 1) as f64 / 100.0;
             size = CelestialBodySize::Giant;
         } else if roll_result <= 386 {
             // Gas supergiant (162.001 to 2000)
-            mass = rng.roll(1, 200000 - (16200 - 1), 16200 - 1) as f32 / 100.0;
+            mass = rng.roll(1, 200000 - (16200 - 1), 16200 - 1) as f64 / 100.0;
             size = CelestialBodySize::Supergiant;
         } else if roll_result > 396 && roll_result <= 400 {
             // TODO: Brown dwarf
-            mass = rng.roll(1, 800000 - (413140 - 1), 413140 - 1) as f32 / 100.0;
+            mass = rng.roll(1, 800000 - (413140 - 1), 413140 - 1) as f64 / 100.0;
             size = CelestialBodySize::Hypergiant;
         } else {
             // Gas hypergiant (2000.001 to 4131)
-            mass = rng.roll(1, 413100 - (200000 - 1), 200000 - 1) as f32 / 100.0;
+            mass = rng.roll(1, 413100 - (200000 - 1), 200000 - 1) as f64 / 100.0;
             size = CelestialBodySize::Hypergiant;
         }
 
         if discriminant(&object) == discriminant(&AstronomicalObject::Void) {
             let density = MathUtils::round_f32_to_precision(
-                interpolate_density(mass) + (rng.roll(1, 61, -31) as f32 / 100.0),
+                (interpolate_density(mass) + (rng.roll(1, 61, -31) as f64 / 100.0)) as f32,
                 4,
             );
-            let radius = calculate_radius(mass as f64, density as f64) as f32;
+            let radius = calculate_radius(mass, density as f64);
             let surface_gravity = calculate_surface_gravity(density, radius);
             // TODO: Atmospheric composition
 
@@ -189,7 +189,7 @@ impl GaseousBodyDetails {
     }
 
     fn get_gas_body_size_modifier(
-        star_mass: f32,
+        star_mass: f64,
         star_type: &StarSpectralType,
         rng: &mut SeededDiceRoller,
         is_proto_giant: bool,
@@ -217,14 +217,14 @@ impl GaseousBodyDetails {
     }
 }
 
-pub fn interpolate_density(mass: f32) -> f32 {
+pub fn interpolate_density(mass: f64) -> f64 {
     let mut prev_point = MASS_TO_DENSITY_DATASET[0];
     for &point in MASS_TO_DENSITY_DATASET.iter().skip(1) {
         if mass <= prev_point.0 && mass >= point.0 {
             if prev_point.0 == point.0 {
                 return point.1;
             }
-            let fraction = (mass - point.0) as f32 / (prev_point.0 - point.0) as f32;
+            let fraction = (mass - point.0) / (prev_point.0 - point.0);
             return point.1 + fraction * (prev_point.1 - point.1);
         }
         prev_point = point;
